requirements.txt
django==3.2.9
black==21.11b1
djangorestframework==3.12.4
django-filter==21.1
jwt==1.3.1
django-environ==0.8.1
